requirements.txt
flask
requests
python-pptx
selenium
webdriver-manager
